ai-fullstack-scaffold/backend/app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.20
psycopg2-binary==2.9.7